Use simple bonus value for now
