+++ style/tab/monster-gear.css
@@ -25,7 +25,7 @@
     text-align: center;
   }
   
-  .forbidden-lands .monster .gears .bonus .skull.damaged {
+  .forbidden-lands .monster .gears .bonus.broken {
     color: #a00;
   }
