+++ style/tab/stronghold-gear.css
@@ -8,7 +8,7 @@
   
   .forbidden-lands .stronghold .gears .name {
     align-items: center;
-    flex-basis: 55%;
+    flex-basis: 60%;
     display: flex;
   }
   
@@ -25,7 +25,7 @@
     text-align: center;
   }
   
-  .forbidden-lands .stronghold .gears .bonus .skull.damaged {
+  .forbidden-lands .stronghold .gears .bonus.broken {
     color: #a00;
   }
   
@@ -35,7 +35,7 @@
   }
   
   .forbidden-lands .stronghold .gears .button {
-    flex-basis: 15%;
+    flex-basis: 10%;
     font-size: 10px;
     text-align: right;
   }
